+++ boomerang/loader/DOS4GWBinaryFile.h
@@ -33,12 +33,12 @@
 // Given a little endian value x, load its value assuming little endian order
 // Note: must be able to take address of x
 // Note: Unlike the LH macro in BinaryFile.h, the paraeter is not a pointer
-#define LMMH(x) ((unsigned)((Byte *)(&x))[0] + ((unsigned)((Byte *)(&x))[1] << 8) + \
-	((unsigned)((Byte *)(&x))[2] << 16) + ((unsigned)((Byte *)(&x))[3] << 24))
+#define LMMH(x) ((uintptr_t)((Byte *)(&x))[0] + ((uintptr_t)((Byte *)(&x))[1] << 8) + \
+	((uintptr_t)((Byte *)(&x))[2] << 16) + ((uintptr_t)((Byte *)(&x))[3] << 24))
 // With this one, x IS a pounsigneder
-#define LMMH2(x) ((unsigned)((Byte *)(x))[0] + ((unsigned)((Byte *)(x))[1] << 8) + \
-	((unsigned)((Byte *)(x))[2] << 16) + ((unsigned)((Byte *)(x))[3] << 24))
-#define LMMHw(x) ((unsigned)((Byte *)(&x))[0] + ((unsigned)((Byte *)(&x))[1] << 8))
+#define LMMH2(x) ((uintptr_t)((Byte *)(x))[0] + ((uintptr_t)((Byte *)(x))[1] << 8) + \
+	((uintptr_t)((Byte *)(x))[2] << 16) + ((uintptr_t)((Byte *)(x))[3] << 24))
+#define LMMHw(x) ((uintptr_t)((Byte *)(&x))[0] + ((uintptr_t)((Byte *)(&x))[1] << 8))
 
 typedef struct { /* exe file header, just the signature really */
     Byte sigLo; /* .EXE signature: 0x4D 0x5A	 */
@@ -171,14 +171,14 @@ public:
 
 protected:
 
-    int dos4gwRead2(short *ps) const; // Read 2 bytes from native addr
-    int dos4gwRead4(int *pi) const; // Read 4 bytes from native addr
+    int16_t dos4gwRead2(int16_t *ps) const; // Read 2 bytes from native addr
+    int32_t dos4gwRead4(int32_t *pi) const; // Read 4 bytes from native addr
 
 public:
 
-    virtual int readNative1(ADDRESS a); // Read 1 bytes from native addr
-    virtual int readNative2(ADDRESS a); // Read 2 bytes from native addr
-    virtual int readNative4(ADDRESS a); // Read 4 bytes from native addr
+    virtual int8_t readNative1(ADDRESS a); // Read 1 bytes from native addr
+    virtual int16_t readNative2(ADDRESS a); // Read 2 bytes from native addr
+    virtual int32_t readNative4(ADDRESS a); // Read 4 bytes from native addr
     virtual QWord readNative8(ADDRESS a); // Read 8 bytes from native addr
     virtual float readNativeFloat4(ADDRESS a); // Read 4 bytes as float
     virtual double readNativeFloat8(ADDRESS a); // Read 8 bytes as float
